:bug: Fix failing test
