+++ package.json
@@ -40,7 +40,7 @@
     "css-mediaquery": "^0.1.2",
     "eventemitter3": "^4.0.7",
     "js-big-decimal": "^1.3.4",
-    "next": "^12.1.0",
+    "next": "^12.1.5",
     "next-i18next": "^10.5.0",
     "next-pwa": "^5.4.4",
     "polished": "^4.1.3",
@@ -60,7 +60,7 @@
     "@testing-library/react": "^13.0.0",
     "babel-plugin-styled-components": "^1.12.0",
     "cross-env": "^7.0.3",
-    "cypress": "9.6.1",
+    "cypress": "^10.1.0",
     "cypress-file-upload": "^5.0.8",
     "cypress-react-selector": "^2.3.16",
     "eslint": "^8.7.0",
